Fix error with signed/unsigned mismatch

+++ PowerEditor/src/NppCommands.cpp
@@ -3845,7 +3845,7 @@ void Notepad_plus::command(int id)
 			{
 				activateDoc(id-IDM_WINDOW_MRU_FIRST);
 			}
-			else if ((id >= IDM_DROPLIST_MRU_FIRST) && (id < (IDM_DROPLIST_MRU_FIRST + _pDocTab->nbItem())))
+			else if ((id >= IDM_DROPLIST_MRU_FIRST) && (id < (IDM_DROPLIST_MRU_FIRST + static_cast<int32_t>(_pDocTab->nbItem()))))
 			{
 				activateDoc(id - IDM_DROPLIST_MRU_FIRST);
 			}
